config: chmod lockfile only when mode differs
Certain filesystems backed by FUSE or on WSL fail
to do chmod. The chmod on the config lockfile is
done to preserve permission mode of the original
lock file. If the lockfile has already the same
mode, there is no need to change it. And on
filesystems without chmod support the mode will be
the same anyway.

Signed-off-by: Konrad Bucheli <konrad.bucheli@psi.ch>